allow host networking for taserver containers

+++ agent/scripts/setup.sh
@@ -5,11 +5,14 @@ dockeruser=${1:-$USER}
 
 # Install docker
 # https://docs.docker.com/engine/install/ubuntu/#install-using-the-convenience-script
-curl -L "https://get.docker.com" -o "get-docker.sh"
-sudo sh get-docker.sh && rm get-docker.sh
-
-# setup current user for docker
-sudo usermod -aG docker $dockeruser
+if [ -z $(which docker) ]; then
+  curl -L "https://get.docker.com" -o "get-docker.sh"
+  sudo sh get-docker.sh && rm get-docker.sh
+  # setup current user for docker
+  sudo usermod -aG docker $dockeruser
+else
+  echo "Docker is already installed, skipping"
+fi
 
 # get taserver image
 docker pull public.ecr.aws/i2q9d4v7/taserver:latest
@@ -17,4 +20,4 @@ docker tag public.ecr.aws/i2q9d4v7/taserver:latest taserver
 
 sudo apt-get update
 sudo apt-get install -y python3 python3-pip 
-pip3 install pyyaml
+pip3 install -r agent/requirements.txt
